src/requirements.txt
Django==1.6.5
bitstring==3.1.3
cffi==0.8.2
crypto==1.1.0
cryptography==0.4
django-appconf==0.6
-e git+https://github.com/maxaon/django-bower.git@b96e882d148a765dd2c1c3903a74f5397830206a#egg=django_bower-master
django-evolution==0.6.9
django-guardian==1.2.0
-e git+https://github.com/maxaon/django-pipeline.git@f74c486b4783be31f789171ceb9653941c4432af#egg=django_pipeline-origin/HEAD
django-security==0.1.22b
djangorestframework==2.3.13
-e git+https://github.com/maxaon/djangular.git@ddb0ccca39a7be3d9870aa67bda2c866165261d9#egg=djangular-master
futures==2.1.6
gpglib==0.1.1
pyasn1==0.1.7
pycparser==2.10
pycrypto==2.6
pygpg==1.1
rsa==3.1.4
six==1.6.1
